{"schema":"decodex/commit/1","summary":"Modularize frozen overlay renderer","authority":"manual"}
